Clean all compile warning.

+++ http/http_conn.cpp
@@ -88,7 +88,7 @@ bool http_conn::write() {
         bytes_to_send -= temp;
         bytes_have_send += temp;
 
-        if (bytes_have_send >= m_iv[0].iov_len) {
+        if (static_cast<unsigned>(bytes_have_send) >= m_iv[0].iov_len) {
             m_iv[0].iov_len = 0;
             m_iv[1].iov_base = response.getFileAddr() + (bytes_have_send - m_write_idx);
             m_iv[1].iov_len = bytes_to_send;
